Moved Modal to it's own directory in components
