--- services/web-app/components/gallery/artistBioModal/index.ts
@@ -1,2 +1,0 @@
-export * from "./artistInfoModalContent"
-export * from "./modal"
