+++ services/web-app/components/modal/index.tsx
@@ -1,0 +1,1 @@
+export * from "./modal";
